requirements.txt
requests==2.31.0
feedparser==6.0.10
beautifulsoup4==4.12.2
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
huggingface-hub==0.20.3
faiss-cpu==1.7.4
streamlit==1.39.0
pandas==2.1.4
plotly==5.18.0
python-dotenv==1.0.0
openai==1.7.2
textblob==0.17.1
newspaper3k==0.2.8
lxml==4.9.3
scikit-learn==1.3.2
numpy==1.24.3
joblib==1.3.2